API: Make plugins a package
