+++ spyder/api/plugin_registration/registry.py
@@ -15,10 +15,9 @@ from qtpy.QtCore import QObject, Signal
 
 # Local imports
 from spyder.config.manager import CONF
-from spyder.api.plugins_enum import Plugins
 from spyder.api.exceptions import SpyderAPIError
 from spyder.api.plugins import (
-    SpyderPluginV2, SpyderDockablePlugin, SpyderPluginWidget,
+    Plugins, SpyderPluginV2, SpyderDockablePlugin, SpyderPluginWidget,
     SpyderPlugin)
